rename the class files to match their classes
Suggestion by @Ecconia
