+++ hints/gui/shopping/shopping_list_tab.py
